+++ processor/src/errors.rs
@@ -6,7 +6,7 @@ use alloc::{string::String, sync::Arc, vec::Vec};
 use miden_air::RowIndex;
 use miden_core::{
     EventId, EventName, Felt, QuadFelt, Word,
-    mast::{DecoratorId, MastForest, MastNodeErrorContext, MastNodeId},
+    mast::{DecoratorId, MastForest, MastNodeId},
     stack::MIN_STACK_DEPTH,
     utils::to_hex,
 };
@@ -563,35 +563,32 @@ pub struct ErrorContextImpl {
 }
 
 impl ErrorContextImpl {
-    pub fn new(
-        mast_forest: &MastForest,
-        node: &impl MastNodeErrorContext,
-        host: &impl BaseHost,
-    ) -> Self {
+    pub fn new(mast_forest: &MastForest, node_id: MastNodeId, host: &impl BaseHost) -> Self {
         let (label, source_file) =
-            Self::precalc_label_and_source_file(None, mast_forest, node, host);
+            Self::precalc_label_and_source_file(None, mast_forest, node_id, host);
         Self { label, source_file }
     }
 
     pub fn new_with_op_idx(
         mast_forest: &MastForest,
-        node: &impl MastNodeErrorContext,
+        node_id: MastNodeId,
         host: &impl BaseHost,
         op_idx: usize,
     ) -> Self {
         let op_idx = op_idx.into();
         let (label, source_file) =
-            Self::precalc_label_and_source_file(op_idx, mast_forest, node, host);
+            Self::precalc_label_and_source_file(op_idx, mast_forest, node_id, host);
         Self { label, source_file }
     }
 
     fn precalc_label_and_source_file(
         op_idx: Option<usize>,
         mast_forest: &MastForest,
-        node: &impl MastNodeErrorContext,
+        node_id: MastNodeId,
         host: &impl BaseHost,
     ) -> (SourceSpan, Option<Arc<SourceFile>>) {
-        node.get_assembly_op(mast_forest, op_idx)
+        mast_forest
+            .get_assembly_op(node_id, op_idx)
             .and_then(|assembly_op| assembly_op.location())
             .map_or_else(
                 || (SourceSpan::UNKNOWN, None),