+++ processor/src/fast/loop.rs
@@ -69,7 +69,7 @@ impl FastProcessor {
             // Execute decorators that should be executed after exiting the node
             self.execute_after_exit_decorators(current_node_id, current_forest, host)?;
         } else {
-            let err_ctx = err_ctx!(current_forest, loop_node, host);
+            let err_ctx = err_ctx!(current_forest, current_node_id, host);
             return Err(ExecutionError::not_binary_value_loop(condition, &err_ctx));
         }
         Ok(())
@@ -121,7 +121,7 @@ impl FastProcessor {
             self.increment_clk(tracer);
             self.execute_after_exit_decorators(current_node_id, current_forest, host)?;
         } else {
-            let err_ctx = err_ctx!(current_forest, loop_node, host);
+            let err_ctx = err_ctx!(current_forest, current_node_id, host);
             return Err(ExecutionError::not_binary_value_loop(condition, &err_ctx));
         }
         Ok(())